update server info

+++ webapp/backend/src/db/entity/Item.js
@@ -6,7 +6,7 @@ module.exports = new EntitySchema({
   columns: {
     id: {
       primary: true,
-      type: "int",
+      type: "bigint",
       generated: true,
     },
     title: {
